+++ elmyra.ip.access.epo/CHANGES.rst
@@ -11,6 +11,7 @@ development
 - ui: display pager on top of and at bottom of resultlist
 - ui: don't show pagers when there are no results yet
 - ui: link to family information (INPADOC, OPS)
+- ui: display cited references below abstract
 
 0.3.0
 =====